refactor: extract matchArrayValue
