+++ src/helpers/matchers/matchArrayValue.ts
@@ -1,0 +1,14 @@
+export function matchArrayValue(
+  prop: Array<string> | undefined,
+  matcher: string | Array<string>
+): boolean {
+  if (!prop || matcher.length === 0) {
+    return false;
+  }
+
+  if (typeof matcher === 'string') {
+    return prop.includes(matcher);
+  }
+
+  return !matcher.some((e) => !prop.includes(e));
+}
