+++ src/queries/a11yStates.ts
@@ -1,5 +1,6 @@
 import type { ReactTestInstance } from 'react-test-renderer';
 import { AccessibilityState } from 'react-native';
+import { matchArrayValue } from '../helpers/matchers/matchArrayValue';
 import { makeQueries } from './makeQueries';
 import type {
   FindAllByQuery,
@@ -14,21 +15,6 @@ type AccessibilityStateKey = keyof AccessibilityState;
 type AccessibilityStateKeys =
   | AccessibilityStateKey
   | Array<AccessibilityStateKey>;
-
-export function matchArrayValue(
-  prop: Array<string> | undefined,
-  matcher: string | Array<string>
-): boolean {
-  if (!prop || matcher.length === 0) {
-    return false;
-  }
-
-  if (typeof matcher === 'string') {
-    return prop.includes(matcher);
-  }
-
-  return !matcher.some((e) => !prop.includes(e));
-}
 
 const queryAllByA11yStates = (
   instance: ReactTestInstance
